tests/Zello.Api.UnitTests/TaskControllerTests.cs
using System.Security.Claims;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using Zello.Api.Controllers;
using Zello.Application.Dtos;
using Zello.Application.Features.Tasks.Models;
using Zello.Domain.Entities;
using Zello.Domain.Entities.Api.User;
using Zello.Domain.Enums;
using Zello.Infrastructure.Data;

namespace Zello.Api.UnitTests;

public class TaskControllerTests : IDisposable {
    private readonly TaskController _controller;
    private readonly ApplicationDbContext _context;
    private readonly Guid _userId;
    private readonly Guid _workspaceId;
    private readonly Guid _projectId;
    private readonly Guid _listId;
    private readonly Guid _taskId;

    public TaskControllerTests() {
        var options = new DbContextOptionsBuilder<ApplicationDbContext>()
            .UseInMemoryDatabase(databaseName: Guid.NewGuid().ToString())
            .Options;
        _context = new ApplicationDbContext(options);
        _controller = new TaskController(_context);

        // Setup user and claims
        _userId = Guid.NewGuid();
        var claims = new List<Claim> {
            new Claim("UserId", _userId.ToString())
        };
        var identity = new ClaimsIdentity(claims);
        var principal = new ClaimsPrincipal(identity);

        _controller.ControllerContext = new ControllerContext {
            HttpContext = new DefaultHttpContext { User = principal }
        };

        // Setup test data
        _workspaceId = Guid.NewGuid();
        _projectId = Guid.NewGuid();
        _listId = Guid.NewGuid();
        _taskId = Guid.NewGuid();
        SetupTestData();
    }

    private void SetupTestData() {
        _context.Users.Add(new User {
            Id = _userId,
            Username = "testuser",
            Name = "Test User",
            Email = "test@example.com",
            PasswordHash = "hashedpassword123",
            CreatedDate = DateTime.UtcNow
        });

        var workspace = new Workspace {
            Id = _workspaceId,
            Name = "Test Workspace",
            OwnerId = _userId,
            CreatedDate = DateTime.UtcNow
        };
        _context.Workspaces.Add(workspace);

        var workspaceMember = new WorkspaceMember {
            Id = Guid.NewGuid(),
            WorkspaceId = _workspaceId,
            UserId = _userId,
            AccessLevel = AccessLevel.Owner,
            CreatedDate = DateTime.UtcNow
        };
        _context.WorkspaceMembers.Add(workspaceMember);

        var project = new Project {
            Id = _projectId,
            WorkspaceId = _workspaceId,
            Name = "Test Project",
            Status = ProjectStatus.InProgress,
            CreatedDate = DateTime.UtcNow
        };
        _context.Projects.Add(project);

        var list = new TaskList {
            Id = _listId,
            ProjectId = _projectId,
            Name = "Test List",
            Position = 0,
            CreatedDate = DateTime.UtcNow
        };
        _context.Lists.Add(list);

        var task = new WorkTask {
            Id = _taskId,
            ProjectId = _projectId,
            ListId = _listId,
            Name = "Test Task",
            Description = "Test Description",
            Status = CurrentTaskStatus.NotStarted,
            Priority = Priority.Medium,
            CreatedDate = DateTime.UtcNow
        };
        _context.Tasks.Add(task);

        _context.SaveChanges();
    }

    [Fact]
    public async Task GetTaskById_ExistingTask_ReturnsTask() {
        // Act
        var result = await _controller.GetTaskById(_taskId);

        // Assert
        var okResult = Assert.IsType<OkObjectResult>(result);
        var taskDto = Assert.IsType<TaskReadDto>(okResult.Value);
        Assert.Equal(_taskId, taskDto.Id);
    }

    [Fact]
    public async Task GetTaskById_NonExistentTask_ReturnsNotFound() {
        // Act
        var result = await _controller.GetTaskById(Guid.NewGuid());

        // Assert
        Assert.IsType<NotFoundObjectResult>(result);
    }

    [Fact]
    public async Task UpdateTask_ValidUpdate_ReturnsUpdatedTask() {
        // Arrange
        var updateDto = new TaskUpdateDto {
            Name = "Updated Task",
            Description = "Updated Description",
            Status = CurrentTaskStatus.InProgress,
            Priority = Priority.High,
            Deadline = DateTime.UtcNow.AddDays(1) // Use Deadline instead
        };

        // Act
        var result = await _controller.UpdateTask(_taskId, updateDto);

        // Assert
        var okResult = Assert.IsType<OkObjectResult>(result);
        var taskDto = Assert.IsType<TaskReadDto>(okResult.Value);
        Assert.Equal(updateDto.Name, taskDto.Name);
        Assert.Equal(updateDto.Status, taskDto.Status);
        Assert.Equal(updateDto.Priority, taskDto.Priority);
    }

    [Fact]
    public async Task MoveTask_ValidMove_ReturnsUpdatedTask() {
        // Arrange
        var newListId = Guid.NewGuid();
        var newList = new TaskList {
            Id = newListId,
            ProjectId = _projectId,
            Name = "New List",
            Position = 1,
            CreatedDate = DateTime.UtcNow
        };
        await _context.Lists.AddAsync(newList);
        await _context.SaveChangesAsync();

        var request = new MoveTaskRequest(newListId);

        try {
            // Act
            var result = await _controller.MoveTask(_taskId, request);

            // Assert
            var okResult = Assert.IsType<OkObjectResult>(result);
            var taskDto = Assert.IsType<TaskReadDto>(okResult.Value);
            Assert.Equal(newListId, taskDto.ListId);
        } catch (InvalidOperationException ex) when (ex.Message.Contains(
                                                         "Transactions are not supported")) {
            // Ignore transaction warnings for in-memory database
        }
    }

    [Fact]
    public async Task AssignUserToTask_ValidAssignment_ReturnsCreated() {
        // Arrange
        var request = new AssignUserRequest { UserId = _userId };

        // Act
        var result = await _controller.AssignUserToTask(_taskId, request);

        // Assert
        var createdResult = Assert.IsType<CreatedAtActionResult>(result);
        var assigneeDto = Assert.IsType<TaskAssigneeReadDto>(createdResult.Value);
        Assert.Equal(_userId, assigneeDto.UserId);
    }

    [Fact]
    public async Task AddTaskComment_ValidComment_ReturnsCreated() {
        // Arrange
        var request = new AddCommentRequest("Test comment content");

        // Act
        var result = await _controller.AddTaskComment(_taskId, request);

        // Assert
        var createdResult = Assert.IsType<CreatedAtActionResult>(result);
        var commentDto = Assert.IsType<CommentReadDto>(createdResult.Value);
        Assert.Equal(request.Content, commentDto.Content);
    }

    [Fact]
    public async Task GetTaskComments_ExistingTask_ReturnsComments() {
        // Arrange
        var comment = new Comment {
            Id = Guid.NewGuid(),
            TaskId = _taskId,
            UserId = _userId,
            Content = "Test Comment",
            CreatedDate = DateTime.UtcNow
        };
        await _context.Comments.AddAsync(comment);
        await _context.SaveChangesAsync();

        // Act
        var result = _controller.GetTaskComments(_taskId);

        // Assert
        var okResult = Assert.IsType<OkObjectResult>(result);
        var comments = Assert.IsType<List<CommentReadDto>>(okResult.Value);
        Assert.Single(comments);
        Assert.Equal(comment.Content, comments[0].Content);
    }

    [Fact]
    public async Task RemoveTaskAssignee_ExistingAssignment_ReturnsNoContent() {
        // Arrange
        var assignee = new TaskAssignee {
            Id = Guid.NewGuid(),
            TaskId = _taskId,
            UserId = _userId,
            AssignedDate = DateTime.UtcNow
        };
        await _context.TaskAssignees.AddAsync(assignee);
        await _context.SaveChangesAsync();

        // Act
        var result = await _controller.RemoveTaskAssignee(_taskId, _userId);

        // Assert
        Assert.IsType<NoContentResult>(result);
    }

    public void Dispose() {
        _context.Database.EnsureDeleted();
        _context.Dispose();
    }
}
